Fix problem with static ICU libraries
ICU requires that the define U_STATIC_IMPLEMENTATION is set for
compiling ICU client code when linking with static ICU libraries
on Windows. The LETypes.h header referenced the internal
cmemory.h header of ICU when the define U_STATIC_IMPLEMENTATION
was set. Therefore compilation of icu-le-hb failed in this case.

Remove the inclusion of ICU's cmemory.h header in general to fix
this problem.